major refactor to modularize
