+++ example/src/pinch.html
@@ -1,0 +1,28 @@
+<!doctype html>
+<head>
+  <title>React-Tappable Example</title>
+  <meta charset="utf-8" />
+  <meta name="format-detection" content="telephone=no" />
+  <meta name="msapplication-tap-highlight" content="no" />
+  <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1, width=device-width, height=device-height, target-densitydpi=device-dpi" />
+  <meta name="apple-mobile-web-app-capable" content="yes" />
+  <meta name="mobile-web-app-capable" content="yes" />
+  <link rel="stylesheet" href="example.css">
+</head>
+<body>
+  <div class="container">
+    <h1>React Tappable</h1>
+    <h2><a href="http://github.com/JedWatson/react-tappable">View project on GitHub</a></h2>
+    <div id="app"></div>
+    <div class="hint">
+      Events in the area on the left will be logged on the right.
+      Toggle scrolling to test event cancel on scroll (touch devices only).
+    </div>
+    <div class="footer">
+      Copyright &copy; Jed Watson 2014. MIT Licensed.
+    </div>
+  </div>
+  <script src="common.js"></script>
+  <script src="bundle.js"></script>
+  <script src="pinch.js"></script>
+</body>
