+++ lib/TapAndPinchable.js
@@ -1,0 +1,18 @@
+'use strict';
+
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
+var TappableMixin = require('./TappableMixin');
+var PinchableMixin = require('./PinchableMixin');
+var getComponent = require('./getComponent');
+var touchStyles = require('./touchStyles');
+
+var Component = getComponent([TappableMixin, PinchableMixin]);
+
+module.exports = Component;
+module.exports.touchStyles = touchStyles;
+module.exports.Mixin = _extends({}, TappableMixin, {
+  onPinchStart: PinchableMixin.onPinchStart,
+  onPinchMove: PinchableMixin.onPinchMove,
+  onPinchEnd: PinchableMixin.onPinchEnd
+});
