+++ lib/TappableMixin.js
@@ -1,6 +1,8 @@
+'use strict';
+
 var React = require('react');
 
-function getTouchProps (touch) {
+function getTouchProps(touch) {
 	if (!touch) return {};
 	return {
 		pageX: touch.pageX,
@@ -12,25 +14,25 @@ function getTouchProps (touch) {
 
 var Mixin = {
 	propTypes: {
-		moveThreshold: React.PropTypes.number,       // pixels to move before cancelling tap
-		activeDelay: React.PropTypes.number,         // ms to wait before adding the `-active` class
-		pressDelay: React.PropTypes.number,          // ms to wait before detecting a press
-		pressMoveThreshold: React.PropTypes.number,  // pixels to move before cancelling press
-		preventDefault: React.PropTypes.bool,        // whether to preventDefault on all events
-		stopPropagation: React.PropTypes.bool,       // whether to stopPropagation on all events
+		moveThreshold: React.PropTypes.number, // pixels to move before cancelling tap
+		activeDelay: React.PropTypes.number, // ms to wait before adding the `-active` class
+		pressDelay: React.PropTypes.number, // ms to wait before detecting a press
+		pressMoveThreshold: React.PropTypes.number, // pixels to move before cancelling press
+		preventDefault: React.PropTypes.bool, // whether to preventDefault on all events
+		stopPropagation: React.PropTypes.bool, // whether to stopPropagation on all events
 
-		onTap: React.PropTypes.func,                 // fires when a tap is detected
-		onPress: React.PropTypes.func,               // fires when a press is detected
-		onTouchStart: React.PropTypes.func,          // pass-through touch event
-		onTouchMove: React.PropTypes.func,           // pass-through touch event
-		onTouchEnd: React.PropTypes.func,            // pass-through touch event
-		onMouseDown: React.PropTypes.func,           // pass-through mouse event
-		onMouseUp: React.PropTypes.func,             // pass-through mouse event
-		onMouseMove: React.PropTypes.func,           // pass-through mouse event
-		onMouseOut: React.PropTypes.func             // pass-through mouse event
+		onTap: React.PropTypes.func, // fires when a tap is detected
+		onPress: React.PropTypes.func, // fires when a press is detected
+		onTouchStart: React.PropTypes.func, // pass-through touch event
+		onTouchMove: React.PropTypes.func, // pass-through touch event
+		onTouchEnd: React.PropTypes.func, // pass-through touch event
+		onMouseDown: React.PropTypes.func, // pass-through mouse event
+		onMouseUp: React.PropTypes.func, // pass-through mouse event
+		onMouseMove: React.PropTypes.func, // pass-through mouse event
+		onMouseOut: React.PropTypes.func // pass-through mouse event
 	},
 
-	getDefaultProps: function () {
+	getDefaultProps: function getDefaultProps() {
 		return {
 			activeDelay: 0,
 			moveThreshold: 100,
@@ -39,7 +41,7 @@ var Mixin = {
 		};
 	},
 
-	getInitialState: function () {
+	getInitialState: function getInitialState() {
 		return {
 			isActive: false,
 			touchActive: false,
@@ -47,18 +49,18 @@ var Mixin = {
 		};
 	},
 
-	componentWillUnmount: function () {
+	componentWillUnmount: function componentWillUnmount() {
 		this.cleanupScrollDetection();
 		this.cancelPressDetection();
 		this.clearActiveTimeout();
 	},
 
-	processEvent: function (event) {
+	processEvent: function processEvent(event) {
 		if (this.props.preventDefault) event.preventDefault();
 		if (this.props.stopPropagation) event.stopPropagation();
 	},
 
-	onTouchStart: function (event) {
+	onTouchStart: function onTouchStart(event) {
 		if (this.props.onTouchStart && this.props.onTouchStart(event) === false) return;
 		this.processEvent(event);
 		window._blockMouseEvents = true;
@@ -67,14 +69,12 @@ var Mixin = {
 			this.initScrollDetection();
 			this.initPressDetection(event, this.endTouch);
 			this._activeTimeout = setTimeout(this.makeActive, this.props.activeDelay);
-		} else if (this.onPinchStart &&
-				(this.props.onPinchStart || this.props.onPinchMove || this.props.onPinchEnd) &&
-				event.touches.length === 2) {
-			this.onPinchStart(event);
+		} else if (this.onPinchStart && (this.props.onPinchStart || this.props.onPinchMove || this.props.onPinchEnd) && event.touches.length === 2) {
+			this.onPinchStart && this.onPinchStart(event);
 		}
 	},
 
-	makeActive: function () {
+	makeActive: function makeActive() {
 		if (!this.isMounted()) return;
 		this.clearActiveTimeout();
 		this.setState({
@@ -82,12 +82,12 @@ var Mixin = {
 		});
 	},
 
-	clearActiveTimeout: function () {
+	clearActiveTimeout: function clearActiveTimeout() {
 		clearTimeout(this._activeTimeout);
 		this._activeTimeout = false;
 	},
 
-	initScrollDetection: function () {
+	initScrollDetection: function initScrollDetection() {
 		this._scrollPos = { top: 0, left: 0 };
 		this._scrollParents = [];
 		this._scrollParentPos = [];
@@ -103,14 +103,14 @@ var Mixin = {
 		}
 	},
 
-	calculateMovement: function (touch) {
+	calculateMovement: function calculateMovement(touch) {
 		return {
 			x: Math.abs(touch.clientX - this._initialTouch.clientX),
 			y: Math.abs(touch.clientY - this._initialTouch.clientY)
 		};
 	},
 
-	detectScroll: function () {
+	detectScroll: function detectScroll() {
 		var currentScrollPos = { top: 0, left: 0 };
 		for (var i = 0; i < this._scrollParents.length; i++) {
 			currentScrollPos.top += this._scrollParents[i].scrollTop;
@@ -119,24 +119,24 @@ var Mixin = {
 		return !(currentScrollPos.top === this._scrollPos.top && currentScrollPos.left === this._scrollPos.left);
 	},
 
-	cleanupScrollDetection: function () {
+	cleanupScrollDetection: function cleanupScrollDetection() {
 		this._scrollParents = undefined;
 		this._scrollPos = undefined;
 	},
 
-	initPressDetection: function (event, callback) {
+	initPressDetection: function initPressDetection(event, callback) {
 		if (!this.props.onPress) return;
-		this._pressTimeout = setTimeout(function () {
+		this._pressTimeout = setTimeout((function () {
 			this.props.onPress(event);
 			callback();
-		}.bind(this), this.props.pressDelay);
+		}).bind(this), this.props.pressDelay);
 	},
 
-	cancelPressDetection: function () {
+	cancelPressDetection: function cancelPressDetection() {
 		clearTimeout(this._pressTimeout);
 	},
 
-	onTouchMove: function (event) {
+	onTouchMove: function onTouchMove(event) {
 		if (this._initialTouch) {
 			this.processEvent(event);
 
@@ -169,7 +169,9 @@ var Mixin = {
 		}
 	},
 
-	onTouchEnd: function (event) {
+	onTouchEnd: function onTouchEnd(event) {
+		var _this = this;
+
 		if (this._initialTouch) {
 			this.processEvent(event);
 			var afterEndTouch;
@@ -177,24 +179,26 @@ var Mixin = {
 			if (movement.x <= this.props.moveThreshold && movement.y <= this.props.moveThreshold && this.props.onTap) {
 				event.preventDefault();
 				event.persist();
-				afterEndTouch = () => {
-					var finalParentScrollPos = this._scrollParents.map(node => node.scrollTop + node.scrollLeft);
-					var stoppedMomentumScroll = this._scrollParentPos.some((end, i) => {
+				afterEndTouch = function () {
+					var finalParentScrollPos = _this._scrollParents.map(function (node) {
+						return node.scrollTop + node.scrollLeft;
+					});
+					var stoppedMomentumScroll = _this._scrollParentPos.some(function (end, i) {
 						return end !== finalParentScrollPos[i];
 					});
 					if (!stoppedMomentumScroll) {
-						this.props.onTap(event);
+						_this.props.onTap(event);
 					}
 				};
 			}
 			this.endTouch(event, afterEndTouch);
-		} else if (this.onPinchEnd && this._initialPinch && (event.touches.length + event.changedTouches.length) === 2) {
+		} else if (this.onPinchEnd && this._initialPinch && event.touches.length + event.changedTouches.length === 2) {
 			this.onPinchEnd(event);
 			event.preventDefault();
 		}
 	},
 
-	endTouch: function (event, callback) {
+	endTouch: function endTouch(event, callback) {
 		this.cancelPressDetection();
 		this.clearActiveTimeout();
 		if (event && this.props.onTouchEnd) {
@@ -211,7 +215,7 @@ var Mixin = {
 		}
 	},
 
-	onMouseDown: function (event) {
+	onMouseDown: function onMouseDown(event) {
 		if (window._blockMouseEvents) {
 			window._blockMouseEvents = false;
 			return;
@@ -225,13 +229,13 @@ var Mixin = {
 		});
 	},
 
-	onMouseMove: function (event) {
+	onMouseMove: function onMouseMove(event) {
 		if (window._blockMouseEvents || !this._mouseDown) return;
 		this.processEvent(event);
 		this.props.onMouseMove && this.props.onMouseMove(event);
 	},
 
-	onMouseUp: function (event) {
+	onMouseUp: function onMouseUp(event) {
 		if (window._blockMouseEvents || !this._mouseDown) return;
 		this.processEvent(event);
 		this.props.onMouseUp && this.props.onMouseUp(event);
@@ -239,14 +243,14 @@ var Mixin = {
 		this.endMouseEvent();
 	},
 
-	onMouseOut: function (event) {
+	onMouseOut: function onMouseOut(event) {
 		if (window._blockMouseEvents || !this._mouseDown) return;
 		this.processEvent(event);
 		this.props.onMouseOut && this.props.onMouseOut(event);
 		this.endMouseEvent();
 	},
 
-	endMouseEvent: function () {
+	endMouseEvent: function endMouseEvent() {
 		this.cancelPressDetection();
 		this._mouseDown = false;
 		this.setState({
@@ -254,12 +258,12 @@ var Mixin = {
 		});
 	},
 
-	cancelTap: function () {
+	cancelTap: function cancelTap() {
 		this.endTouch();
 		this._mouseDown = false;
 	},
 
-	handlers: function () {
+	handlers: function handlers() {
 		return {
 			onTouchStart: this.onTouchStart,
 			onTouchMove: this.onTouchMove,
