+++ lib/touchStyles.js
@@ -1,0 +1,14 @@
+'use strict';
+
+var touchStyles = {
+  WebkitTapHighlightColor: 'rgba(0,0,0,0)',
+  WebkitTouchCallout: 'none',
+  WebkitUserSelect: 'none',
+  KhtmlUserSelect: 'none',
+  MozUserSelect: 'none',
+  msUserSelect: 'none',
+  userSelect: 'none',
+  cursor: 'pointer'
+};
+
+module.exports = touchStyles;
