+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-tappable",
   "version": "0.5.2",
   "description": "Touch / Tappable Event Handling Component for React",
-  "main": "lib/Tappable.js",
+  "main": "lib/TapAndPinchable.js",
   "author": "Jed Watson",
   "license": "MIT",
   "repository": {
