+++ src/Pinchable.js
@@ -1,0 +1,7 @@
+var PinchableBaseMixin = require('./PinchableBaseMixin');
+var PinchableMixin = require('./PinchableMixin');
+var getComponent = require('./getComponent');
+
+var Component = getComponent([PinchableBaseMixin, PinchableMixin]);
+
+module.exports = Component;
